requirements.txt
streamlit
pandas
numpy
xgboost>=2.1.0
shap
matplotlib
scikit-learn
langchain-core>=0.3,<0.4
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
openai
python-dotenv
tiktoken


prefect>=2.14.0
duckdb>=0.9.0
dbt-core>=1.7.0
dbt-duckdb>=1.7.0
pytest>=7.0.0
boto3>=1.28.0
fastapi
uvicorn[standard]